XP-695 Move XmlMixinParser to core-schema module
- Moved XmlMixinParser to core-schema module. Moved it under com.enonic.xp.core.impl.schema.mixin package. Also moved the associated tests.
