+++ modules/core-schema/src/main/java/com/enonic/xp/core/impl/schema/mixin/MixinLoader.java
@@ -20,7 +20,6 @@ import com.enonic.xp.module.ModuleKey;
 import com.enonic.xp.schema.mixin.Mixin;
 import com.enonic.xp.schema.mixin.MixinName;
 import com.enonic.xp.schema.mixin.Mixins;
-import com.enonic.xp.xml.parser.XmlMixinParser;
 
 final class MixinLoader
 {
