+++ modules/core-schema/src/main/java/com/enonic/xp/core/impl/schema/mixin/XmlMixinParser.java
@@ -1,10 +1,12 @@
-package com.enonic.xp.xml.parser;
+package com.enonic.xp.core.impl.schema.mixin;
 
 import com.google.common.annotations.Beta;
 
 import com.enonic.xp.form.Form;
 import com.enonic.xp.schema.mixin.Mixin;
 import com.enonic.xp.xml.DomElement;
+import com.enonic.xp.xml.parser.XmlFormMapper;
+import com.enonic.xp.xml.parser.XmlModelParser;
 
 @Beta
 public final class XmlMixinParser
